Handle command-line mods/maps/demos/savefiles
If a single file or directory is passed on the command line
we try to infer the basedir, mod, and quake flavor (original
or remastered) as needed, and take the appropriate action
based on the type of the argument.

This enables associating the executable with .bsp/.dem/.sav
files and dropping files/mod dirs over the exe in order to
play them.