Narrow down megahealth decay autosave check
Previously, the 3 second damage window that disabled autosaving
was not triggered if the player's health was >100 in order to
avoid counting megahealth decay as taking damage. This meant,
however, that stepping in slime/lava while health was >100
had no effect on autosave.

Now autosave is disabled if the frame-to-frame health delta is
more than 3 hp (which allows for stacking 3 MHs), health is
below 100 or the player is standing in slime/lava.